requirements.txt
streamlit
opencv-python
numpy
Pillow
supabase
python-dotenv
ultralytics
